Remove warnings and add missing noalias

+++ include/hpp/fcl/internal/traversal_node_setup.h
@@ -313,12 +313,12 @@ bool initialize(MeshShapeCollisionTraversalNode<BV, S>& node,
 
   if(!tf1.isIdentity())
   {
-    std::vector<Vec3f> vertices_transformed((size_t)model1.num_vertices);
-    for(int i = 0; i < model1.num_vertices; ++i)
+    std::vector<Vec3f> vertices_transformed(model1.num_vertices);
+    for(unsigned int i = 0; i < model1.num_vertices; ++i)
     {
       const Vec3f & p = model1.vertices[i];
       Vec3f new_v = tf1.transform(p);
-      vertices_transformed[(size_t)i] = new_v;
+      vertices_transformed[i] = new_v;
     }
 
     model1.beginReplaceModel();
@@ -419,11 +419,11 @@ bool initialize(MeshCollisionTraversalNode<BV, RelativeTransformationIsIdentity>
   if(!tf1.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed1(model1.num_vertices);
-    for(int i = 0; i < model1.num_vertices; ++i)
+    for(unsigned int i = 0; i < model1.num_vertices; ++i)
     {
       Vec3f& p = model1.vertices[i];
       Vec3f new_v = tf1.transform(p);
-      vertices_transformed1[(size_t)i] = new_v;
+      vertices_transformed1[i] = new_v;
     }
 
     model1.beginReplaceModel();
@@ -436,11 +436,11 @@ bool initialize(MeshCollisionTraversalNode<BV, RelativeTransformationIsIdentity>
   if(!tf2.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed2(model2.num_vertices);
-    for(int i = 0; i < model2.num_vertices; ++i)
+    for(unsigned int i = 0; i < model2.num_vertices; ++i)
     {
       Vec3f& p = model2.vertices[i];
       Vec3f new_v = tf2.transform(p);
-      vertices_transformed2[(size_t)i] = new_v;
+      vertices_transformed2[i] = new_v;
     }
 
     model2.beginReplaceModel();
@@ -534,11 +534,11 @@ bool initialize(MeshDistanceTraversalNode<BV, RelativeTransformationIsIdentity>&
   if(!tf1.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed1(model1.num_vertices);
-    for(int i = 0; i < model1.num_vertices; ++i)
+    for(unsigned int i = 0; i < model1.num_vertices; ++i)
     {
       const Vec3f & p = model1.vertices[i];
       Vec3f new_v = tf1.transform(p);
-      vertices_transformed1[(size_t)i] = new_v;
+      vertices_transformed1[i] = new_v;
     }
 
     model1.beginReplaceModel();
@@ -551,11 +551,11 @@ bool initialize(MeshDistanceTraversalNode<BV, RelativeTransformationIsIdentity>&
   if(!tf2.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed2(model2.num_vertices);
-    for(int i = 0; i < model2.num_vertices; ++i)
+    for(unsigned int i = 0; i < model2.num_vertices; ++i)
     {
       const Vec3f & p = model2.vertices[i];
       Vec3f new_v = tf2.transform(p);
-      vertices_transformed2[(size_t)i] = new_v;
+      vertices_transformed2[i] = new_v;
     }
 
     model2.beginReplaceModel();
@@ -632,11 +632,11 @@ bool initialize(MeshShapeDistanceTraversalNode<BV, S>& node,
   if(!tf1.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed1(model1.num_vertices);
-    for(int i = 0; i < model1.num_vertices; ++i)
+    for(unsigned int i = 0; i < model1.num_vertices; ++i)
     {
       const Vec3f & p = model1.vertices[i];
       Vec3f new_v = tf1.transform(p);
-      vertices_transformed1[(size_t)i] = new_v;
+      vertices_transformed1[i] = new_v;
     }
 
     model1.beginReplaceModel();
@@ -679,11 +679,11 @@ bool initialize(ShapeMeshDistanceTraversalNode<S, BV>& node,
   if(!tf2.isIdentity())
   {
     std::vector<Vec3f> vertices_transformed(model2.num_vertices);
-    for(int i = 0; i < model2.num_vertices; ++i)
+    for(unsigned int i = 0; i < model2.num_vertices; ++i)
     {
       const Vec3f & p = model2.vertices[i];
       Vec3f new_v = tf2.transform(p);
-      vertices_transformed[(size_t)i] = new_v;
+      vertices_transformed[i] = new_v;
     }
 
     model2.beginReplaceModel();
